Incremented SDK_REVISION. Copied the changes from VerifyQuickSample/verifySDK to the standalone folder verifySDK

+++ verifySDK/src/main/java/com/nexmo/sdk/core/config/Defaults.java
@@ -18,7 +18,7 @@ package com.nexmo.sdk.core.config;
  */
 public class Defaults {
 
-    public static final int MAX_ALLOWABLE_TIME_DELTA = 5 * 60 * 1000;
+    public static final long MAX_ALLOWABLE_TIME_DELTA = 5 * 60 * 1000;
     public static final int CONNECTION_TIMEOUT = 15 * 1000;
     public static final int CONNECTION_READ_TIMEOUT = 10 * 1000;
     public static final int MIN_CODE_LENGTH = 4;
